ai-professor-main/requirements.txt
streamlit==1.41.1
google-generativeai==0.8.4
requests==2.32.3
graphviz==0.20.3
python-dotenv==1.0.1